added chat service with mongodb and socketio(beanio)

+++ api/actions/events.py
@@ -1,5 +1,5 @@
 from db.dals import EventsDAL
-from db.models import EventsOrm
+from db.models.models import EventsOrm
 from sqlalchemy.dialects.postgresql import UUID
 from api.schemas import EventAddDTO, EventShowDTO
 from typing import Optional
@@ -13,6 +13,11 @@ from pydantic import TypeAdapter
 
 redis_client = redis_async.Redis(host="localhost", port=6379, db=0, decode_responses=True)
 
+async def redis_available():
+    try:
+        return await redis_client.ping()
+    except:
+        return False
 
 def orm_to_dict(obj: EventsOrm):
     dict1 = {}
@@ -37,10 +42,10 @@ async def _create_new_event(cred, session) -> EventShowDTO:
         )
 
     event_show_dto = EventShowDTO.model_validate(created_event_orm, from_attributes=True)
-
-    keys = await redis_client.keys("events:page:*")
-    if keys:
-        await redis_client.delete(*keys)
+    if await redis_available():
+        keys = await redis_client.keys("events:page:*")
+        if keys:
+            await redis_client.delete(*keys)
     return event_show_dto
 
 
@@ -61,26 +66,26 @@ async def _get_events_limit_10_by_page(page: int, session) -> list[EventsOrm]:
     cache_key = f"events:page:{page}"
 
     try:
-        cached_data = await redis_client.get(cache_key)
-        
-        if cached_data:
-            hit += 1
-            events_dto = adapter.validate_json(cached_data)
-            print(hit, miss)
-            return [EventsOrm(**dto.model_dump()) for dto in events_dto]
+        if await redis_available():
+
+            cached_data = await redis_client.get(cache_key)
             
+            if cached_data:
+                hit += 1
+                events_dto = adapter.validate_json(cached_data)
+                return [EventsOrm(**dto.model_dump()) for dto in events_dto]
+                
     except Exception as e:
         print(f"Redis error (read): {e}")
     miss += 1
-    print(hit, miss, hit/miss)
     start = 10 * (page - 1)
     events_dal = EventsDAL(session)
     events_orm_list = await events_dal.get_events_limit_10(offset=start)
     
     if not events_orm_list:
         return []
-
-    asyncio.create_task(cache_page_data(cache_key, events_orm_list))
+    if await redis_available():
+        asyncio.create_task(cache_page_data(cache_key, events_orm_list))
 
     return events_orm_list
 
@@ -94,8 +99,6 @@ async def cache_page_data(key: str, events_orm: list[EventsOrm]):
         
     except Exception as e:
         print(f"Redis error (write): {e}")
-
-
 
 async def _get_event_by_id(event_id, session) -> EventsOrm:
     async with session.begin():
